# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATLAS \u2014 Market Intelligence Engine",
    "description": "Jhaveri Intelligence Platform \u2014 Market \u2192 Sector \u2192 Stock \u2192 Decision",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/mf/universe": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Universe",
        "operationId": "get_universe_api_v1_mf_universe_get",
        "parameters": [
          {
            "name": "benchmark",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Benchmark"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "broad_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Broad Category"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniverseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mf/flows": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Flows",
        "operationId": "get_flows_api_v1_mf_flows_get",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 12,
              "title": "Months"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/overlap": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Overlap",
        "operationId": "get_overlap_api_v1_mf_overlap_get",
        "parameters": [
          {
            "name": "funds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated mstar_ids: A,B",
              "title": "Funds"
            },
            "description": "Comma-separated mstar_ids: A,B"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverlapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/holding-stock/{symbol}": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Holding Stock",
        "operationId": "get_holding_stock_api_v1_mf_holding_stock__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingStockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}/holdings": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Holdings",
        "operationId": "get_fund_holdings_api_v1_mf__mstar_id__holdings_get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}/sectors": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Sectors",
        "operationId": "get_fund_sectors_api_v1_mf__mstar_id__sectors_get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundSectorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}/rs-history": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Rs History",
        "operationId": "get_fund_rs_history_api_v1_mf__mstar_id__rs_history_get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          },
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 120,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 12,
              "title": "Months"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundRSHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}/weighted-technicals": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Weighted Technicals",
        "operationId": "get_fund_weighted_technicals_api_v1_mf__mstar_id__weighted_technicals_get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightedTechnicalsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}/nav-history": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Nav History",
        "operationId": "get_fund_nav_history_api_v1_mf__mstar_id__nav_history_get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NAVHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mf/{mstar_id}": {
      "get": {
        "tags": [
          "mf"
        ],
        "summary": "Get Fund Deep Dive",
        "operationId": "get_fund_deep_dive_api_v1_mf__mstar_id__get",
        "parameters": [
          {
            "name": "mstar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mstar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundDeepDiveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BroadCategoryGroup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryGroup"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BroadCategoryGroup"
      },
...
      "FlowRow": {
        "properties": {
          "month_date": {
            "type": "string",
            "format": "date",
            "title": "Month Date"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "net_flow_cr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Flow Cr"
          },
          "gross_inflow_cr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Inflow Cr"
          },
          "gross_outflow_cr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Outflow Cr"
          },
          "aum_cr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aum Cr"
          },
          "sip_flow_cr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sip Flow Cr"
          },
          "sip_accounts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sip Accounts"
          },
          "folios": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folios"
          }
        },
        "type": "object",
        "required": [
          "month_date",
          "category"
        ],
        "title": "FlowRow"
      },
      "FlowsResponse": {
        "properties": {
          "flows": {
            "items": {
              "$ref": "#/components/schemas/FlowRow"
            },
            "type": "array",
            "title": "Flows",
            "default": []
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "data_as_of",
          "staleness"
        ],
        "title": "FlowsResponse"
      },
...
      "FundDeepDiveResponse": {
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/FundIdentity"
          },
          "daily": {
            "$ref": "#/components/schemas/FundDailyMetrics"
          },
          "pillars": {
            "$ref": "#/components/schemas/ConvictionPillarsMF"
          },
          "sector_exposure": {
            "$ref": "#/components/schemas/SectorExposureSummary"
          },
          "top_holdings": {
            "items": {
              "$ref": "#/components/schemas/TopHoldingSummary"
            },
            "type": "array",
            "title": "Top Holdings",
            "default": []
          },
          "weighted_technicals": {
            "$ref": "#/components/schemas/WeightedTechnicalsSummary"
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          },
          "inactive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive"
          },
          "mf_lifecycle_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MFLifecycleEvent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "identity",
          "daily",
          "pillars",
          "sector_exposure",
          "weighted_technicals",
          "data_as_of",
          "staleness"
        ],
        "title": "FundDeepDiveResponse"
      },
      "FundHoldingStockEntry": {
        "properties": {
          "mstar_id": {
            "type": "string",
            "title": "Mstar Id"
          },
          "fund_name": {
            "type": "string",
            "title": "Fund Name"
          },
          "weight_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight Pct"
          }
        },
        "type": "object",
        "required": [
          "mstar_id",
          "fund_name",
          "weight_pct"
        ],
        "title": "FundHoldingStockEntry"
      },
...
      "FundRSHistoryResponse": {
        "properties": {
          "mstar_id": {
            "type": "string",
            "title": "Mstar Id"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RSHistoryPoint"
            },
            "type": "array",
            "title": "Points",
            "default": []
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "mstar_id",
          "data_as_of",
          "staleness"
        ],
        "title": "FundRSHistoryResponse"
      },
      "FundSector": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "weight_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight Pct"
          },
          "stock_count": {
            "type": "integer",
            "title": "Stock Count",
            "default": 0
          },
          "sector_rs_composite": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Rs Composite"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "weight_pct"
        ],
        "title": "FundSector"
      },
      "FundSectorsResponse": {
        "properties": {
          "sectors": {
            "items": {
              "$ref": "#/components/schemas/FundSector"
            },
            "type": "array",
            "title": "Sectors",
            "default": []
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          }
        },
        "type": "object",
        "required": [
          "as_of_date"
        ],
        "title": "FundSectorsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingStockResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "instrument_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Id"
          },
          "funds": {
            "items": {
              "$ref": "#/components/schemas/FundHoldingStockEntry"
            },
            "type": "array",
            "title": "Funds",
            "default": []
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "data_as_of",
          "staleness"
        ],
        "title": "HoldingStockResponse"
      },
      "HoldingsResponse": {
        "properties": {
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/Holding"
            },
            "type": "array",
            "title": "Holdings",
            "default": []
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "coverage_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Coverage Pct"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "as_of_date",
          "coverage_pct"
        ],
        "title": "HoldingsResponse"
      },
...
      "NAVHistoryResponse": {
        "properties": {
          "mstar_id": {
            "type": "string",
            "title": "Mstar Id"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/NAVPoint"
            },
            "type": "array",
            "title": "Points",
            "default": []
          },
          "coverage_gap_days": {
            "type": "integer",
            "title": "Coverage Gap Days",
            "default": 0
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "mstar_id",
          "data_as_of",
          "staleness"
        ],
        "title": "NAVHistoryResponse"
      },
...
      "OverlapHolding": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "weight_a": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight A"
          },
          "weight_b": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight B"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "symbol",
          "weight_a",
          "weight_b"
        ],
        "title": "OverlapHolding"
      },
      "OverlapResponse": {
        "properties": {
          "fund_a": {
            "type": "string",
            "title": "Fund A"
          },
          "fund_b": {
            "type": "string",
            "title": "Fund B"
          },
          "overlap_pct": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Overlap Pct"
          },
          "common_holdings": {
            "items": {
              "$ref": "#/components/schemas/OverlapHolding"
            },
            "type": "array",
            "title": "Common Holdings",
            "default": []
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "fund_a",
          "fund_b",
          "overlap_pct",
          "data_as_of",
          "staleness"
        ],
        "title": "OverlapResponse"
      },
...
      "RSHistoryPoint": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "rs_composite": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rs Composite"
          }
        },
        "type": "object",
        "required": [
          "as_of_date",
          "rs_composite"
        ],
        "title": "RSHistoryPoint"
      },
...
      "Staleness": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "age_minutes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Age Minutes"
          },
          "flag": {
            "$ref": "#/components/schemas/StalenessFlag",
            "default": "FRESH"
          }
        },
        "type": "object",
        "required": [
          "source",
          "age_minutes"
        ],
        "title": "Staleness"
      },
      "StalenessFlag": {
        "type": "string",
        "enum": [
          "FRESH",
          "STALE",
          "EXPIRED"
        ],
        "title": "StalenessFlag"
      },
...
      "UniverseResponse": {
        "properties": {
          "broad_categories": {
            "items": {
              "$ref": "#/components/schemas/BroadCategoryGroup"
            },
            "type": "array",
            "title": "Broad Categories",
            "default": []
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "data_as_of",
          "staleness"
        ],
        "title": "UniverseResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeightedTechnicalsResponse": {
        "properties": {
          "mstar_id": {
            "type": "string",
            "title": "Mstar Id"
          },
          "weighted_rsi": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Rsi"
          },
          "weighted_breadth_pct_above_200dma": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Breadth Pct Above 200Dma"
          },
          "weighted_macd_bullish_pct": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Macd Bullish Pct"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "data_as_of": {
            "type": "string",
            "format": "date",
            "title": "Data As Of"
          },
          "staleness": {
            "$ref": "#/components/schemas/Staleness"
          }
        },
        "type": "object",
        "required": [
          "mstar_id",
          "as_of_date",
          "data_as_of",
          "staleness"
        ],
        "title": "WeightedTechnicalsResponse"
      },
...
    }
  }
}